+++ controller/otp_controller.js
@@ -37,8 +37,7 @@ const { validatePhoneNumber } = require('../utils/validators');
             { upsert: true, new: true, setDefaultsOnInsert: true }
         );
 
-console.log("hello");
-          sendGeneralResponse(res, true, "OTP sent successfully",  200,  {
+           sendGeneralResponse(res, true, "OTP sent successfully",  200,  {
             otp:otp,
          })
         
@@ -48,50 +47,4 @@ console.log("hello");
         sendGeneralResponse(res, false, "Error in sending OTP",  500,)
      }
 };
-
-// // verify OTP
-// const verifyotp = async (req, res) => {
-//     try {
-//         const { phone, otp } = req.body;
-
-//         if (!phone || !otp) {
-//             res.status(400).json({
-//                 success: false, msg: 'Phone number and OTP are required', data:[]})
-//          }
-
-//         // Find the OTP record for the provided phone number
-//         const otpRecord = await Otp.findOne({ phone });
-
-//         if (!otpRecord) {
-         
-//             res.status(400).json({
-//                 success: false, msg: 'Phone number not found', data:[]})
-//         }
-
-//         const { otp: storedOtp, otpExpiration } = otpRecord;
-
-//         // Check if the OTP matches
-//         if (storedOtp !== otp) {
-//             res.status(400).json({
-//                 success: false, msg: 'Incorrect OTP', data:[]})
-//         }
-
-//         // Check if the OTP has expired
-//         if (new Date() > otpExpiration) {
-//             res.status(400).json({
-//                 success: false, msg: 'OTP expired', data:[]})
-           
-//         }
-
-//         // Respond with success message
-//         return { success: true, status: 200, msg: 'OTP verified successfully' };
-
-//     } catch (error) {
-//         console.error('Error in verifying OTP:', error);
-
-//         // Respond with an error message
-//         return { success: false, status: 500, msg: 'Error in verifying OTP' };
-//     }
-// };
-
 module.exports = { sendotp };
