+++ utils/validators.js
@@ -1,25 +1,6 @@
-// /**
-//  * Validate email format
-//  * @param {string} email
-//  * @returns {boolean}
-//  */
-// const validateEmail = (email) => {
-//     const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-//     return emailRegex.test(email);
-// };
 
-// /**
-//  * Validate phone number format
-//  * @param {string} phone
-//  * @returns {boolean}
-//  */
-// const validatePhoneNumber = (phone) => {
-//     const phoneRegex = /^[0-9]{10}$/; // Adjust regex as needed
-//     return phoneRegex.test(phone);
-// };
+const { sendGeneralResponse } = require('./responseHelper');
 
-// module.exports = { validateEmail, validatePhoneNumber };
-// Validator for email
 const validateEmail = (email) => {
     const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
     return emailRegex.test(email);
@@ -27,17 +8,24 @@ const validateEmail = (email) => {
 
 // Validator for phone number
 const validatePhoneNumber = (phone) => {
-    // const phoneRegex = /^\d{10}$/;
     const phoneRegex = /^\+?[1-9]\d{1,14}$/;
 
     return phoneRegex.test(phone);
 };
 
-
-
+// validate empety fields
+function validateRequiredFields(res, fields) {
+     for (const [key, value] of Object.entries(fields)) {
+        if (!value) {
+            return sendGeneralResponse(res, false, `${key} field is required`, 400);
+        }
+    }
+    return true;
+}
 
 module.exports = {
     validateEmail,
-    validatePhoneNumber
+    validatePhoneNumber,
+    validateRequiredFields
 };
 9
